requirements.txt
backports.ssl-match-hostname==3.7.0.1
bcrypt==3.2.0
cachetools==4.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
Click==8.0.1
colorama==0.4.4
configobj==5.0.6
croniter==1.0.15
dnspython==2.1.0
dpath==2.0.1
enum34==1.1.10
Flask==2.0.1
Flask-HTTPAuth==4.4.0
gevent==21.8.0
greenlet==1.1.1
gunicorn==20.1.0
html5lib==1.1
idna==3.2
ipaddress==1.0.23
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
parse-it==2021.8.11.10.10
ply==3.11
pycparser==2.20
pyhcl==0.4.4
pymongo==3.12.0
python-dateutil==2.8.2
python-dotenv==0.19.0
#PyYAML==5.4.1
requests==2.26.0
retrying==1.3.3
six==1.16.0
toml==0.10.2
urllib3==1.26.6
webencodings==0.5.1
websocket-client==1.2.1
Werkzeug==2.0.1
xmltodict==0.12.0
py-machineid==0.5.1
